refactor: products fetch 로직 hook으로 분리
